This commit was manufactured by cvs2svn to create tag 'CHOCOLATE_DOOM_0_1_4'.

+++ chocolate-doom/src/m_swap.h
@@ -31,15 +31,15 @@
 #define __M_SWAP__
 
 
-
+#include "config.h"
 
 // Endianess handling.
 // WAD files are stored little endian.
 #ifdef WORDS_BIGENDIAN
 short	SwapSHORT(short);
-long	SwapLONG(long);
+int	SwapLONG(int);
 #define SHORT(x)	((short)SwapSHORT((unsigned short) (x)))
-#define LONG(x)         ((long)SwapLONG((unsigned long) (x)))
+#define LONG(x)         ((int)SwapLONG((unsigned int) (x)))
 #else
 #define SHORT(x)	(x)
 #define LONG(x)         (x)
@@ -52,6 +52,12 @@ long	SwapLONG(long);
 //-----------------------------------------------------------------------------
 //
 // $Log$
+// Revision 1.5.2.2  2006/01/24 01:47:30  fraggle
+// More endianness fixes
+//
+// Revision 1.5.2.1  2006/01/23 21:56:19  fraggle
+// Include the config header so that endianness is dealt with correctly
+//
 // Revision 1.5  2006/01/05 02:48:03  fraggle
 // Fixes for big endian machines (thanks locust)
 //
